Final step to break out geometry bindings
This moves SceneGraph, SceneGraphInspector, QueryObject, and the return
types of queries into its own compilation and unit tests.

Notably, this "restores" the file `geometry_py.cc`, but blame on the file
has been completely lost. This simplified the last step (keeping master
saner) and all of the original lines of geometry_py.cc have now been
spread around the various geometry_py_*.cc files, so that history isn't
lost.
